tests/wpt/tests/css/css-conditional/container-queries/scroll-state/at-container-scroll-direction-serialization.html

<!DOCTYPE html>
<title>CSS Container Queries: scroll-state(direction) conditionText serialization</title>
<link rel="help" href="https://drafts.csswg.org/css-conditional-5/#container-queries">
<link rel="help" href="https://drafts.csswg.org/cssom/#serialize-a-css-rule">
<link rel="help" href="https://drafts.csswg.org/css-conditional-5/#direction">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/css-conditional/container-queries/support/cq-testcommon.js"></script>
<style id="testSheet">
  @container scroll-state(   direction:top   ) { }
  @container scroll-STate(direction:    ) { }
  @container scroll-STate(direction) { }
  @container  scroll-state(  ( direction: LEFT) OR ( direction: BOTTOM  ) ) { }
  @container scroll-state (direction: right) { }
</style>
<script>
  setup(() => {
    assert_implements_scroll_state_container_queries();
    assert_equals(testSheet.sheet.cssRules.length, 5);
  });

  const tests = [
      ["scroll-state(direction: top)", "Normalize spaces"],
      ["scroll-STate(direction:    )", "No value - invalid, serializes as <general-enclosed>"],
      ["scroll-state(direction)", "Boolean context"],
      ["scroll-state((direction: left) or (direction: bottom))", "Logical with 'or'"],
      ["scroll-state (direction: right)", "Not a scroll-state function with space before '('"]
  ].map((e, i) => [testSheet.sheet.cssRules[i], ...e]);

  tests.forEach((t) => {
    test(() => assert_equals(t[0].conditionText, t[1]), t[2]);
  });
</script>
